+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - [added] Implemented `messaging.MulticastMessage` type and the
   `messaging.SendMulticast()` function for sending the same
   message to multiple recipients.
-- [added] Implementd `messaging.SendAllDryRun()` and
+- [added] Implemented `messaging.SendAllDryRun()` and
   `messaging.SendMulticastDryRun()` functions for sending messages
   in the validate only mode.
 - [added] Implemented `messaging.SendAll()` function for sending